add useForwardUnhandledRequests to server boilerplate
