+++ packages/lesmis-server/src/app/createApp.ts
@@ -2,12 +2,13 @@
  * Tupaia
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  */
-import { IncomingMessage, ServerResponse } from 'http';
 import { TupaiaDatabase } from '@tupaia/database';
 import { LesmisSessionModel } from '../models';
-import { createProxyMiddleware, fixRequestBody } from 'http-proxy-middleware';
-import { Express } from 'express';
-import { OrchestratorApiBuilder, handleWith, handleError } from '@tupaia/server-boilerplate';
+import {
+  OrchestratorApiBuilder,
+  handleWith,
+  useForwardUnhandledRequests,
+} from '@tupaia/server-boilerplate';
 import {
   DashboardRoute,
   EntityRoute,
@@ -22,56 +23,7 @@ import {
 import { attachSession } from '../session';
 import { hasLesmisAccess } from '../utils';
 
-import { Request, Response, NextFunction } from 'express';
-
-import { UnauthenticatedError } from '@tupaia/utils';
-
-export const attachAuthorizationHeader = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  const { session } = req;
-
-  if (!session) {
-    throw new UnauthenticatedError('Session is not attached');
-  }
-
-  req.headers.authorization = await session.getAuthHeader();
-
-  next();
-};
-
-const useForwardUnhandledRequestsToMeditrak = (app: Express, prefix: string) => {
-  const { MEDITRAK_API_URL = 'http://localhost:8090/v2' } = process.env;
-
-  const options = {
-    target: MEDITRAK_API_URL,
-    changeOrigin: true,
-    pathRewrite: (path: string) => {
-      if (prefix && path.startsWith(prefix)) {
-        path = path.replace(prefix, '');
-      }
-
-      // Remove the version string because version is already included in Meditrak base url.
-      if (path.startsWith('/v')) {
-        const secondSlashIndex = path.indexOf('/', 2);
-        const version = parseFloat(path.substring(2, secondSlashIndex));
-        return path.replace(`/v${version}`, '');
-      }
-      return path;
-    },
-    onProxyReq: fixRequestBody,
-    onProxyRes: (proxyRes: IncomingMessage, req: IncomingMessage, res: ServerResponse) => {
-      // To get around CORS because Admin Panel has credentials: true in fetch for session cookies
-      // eslint-disable-next-line no-param-reassign
-      proxyRes.headers['Access-Control-Allow-Origin'] = res.get('Access-Control-Allow-Origin');
-    },
-  };
-
-  // Forward any unhandled request to meditrak-server
-  app.use(attachSession, attachAuthorizationHeader, createProxyMiddleware(options), handleError);
-};
+const { MEDITRAK_API_URL = 'http://localhost:8090/v2' } = process.env;
 
 /**
  * Set up express server with middleware,
@@ -94,7 +46,7 @@ export function createApp() {
     .build();
 
   // Forward any unhandled request to meditrak-server
-  useForwardUnhandledRequestsToMeditrak(app, '/admin');
+  useForwardUnhandledRequests(app, MEDITRAK_API_URL, '/admin', attachSession);
 
   return app;
 }
